requirements.txt
azure-storage-blob
azure-identity
python-dotenv
colorama
